Adds initial release of Flutter Experiment SDK
This commit introduces the initial release of the Amplitude Experiment Flutter SDK.

It includes support for iOS, Android, and Web platforms, along with feature flagging and A/B testing capabilities. Remote evaluation with automatic retries and local storage for variant persistence are also implemented. Comprehensive documentation and an example application are included.